perf(vectorstore): flatten HNSW segment graph to flat CSR + uint32 neighbor ids (#90)
Two-stage HNSW segment-graph topology memory optimization (cross-validated).

Stage 1: loaded read path []map[int][]uint64 -> flat heap CSR (csrNodeBase/csrLayerStart/csrPool), new VSG2 graph-file format. -56.9% topology heap, byte-identical search path.
Stage 2: neighbor ids u64 -> u32 everywhere edges live; getNeighborsRef returns []uint32 (free uint64 cast in the 2 search loops), zero-copy preserved. -77.5% (4.45x) topology vs the original map, 18 allocs/op unchanged, ns/op equal-or-faster, recall identical.
Also: memGraphStore (test-only reference impl) moved into the test build; dead appendU64 removed.

Designed, implemented and adversarially reviewed via multi-agent workflows.

+++ core/vectorstore/graphstore_test.go
@@ -221,7 +221,7 @@ func TestSegGraphStore_MemEquivalence(t *testing.T) {
 // (out-of-range id / nil-neighbors node) that the search tests never hit, since
 // search only ever passes valid node ids.
 func TestSegGraphStore_getNeighborsRef_Guard(t *testing.T) {
-	g := &segGraphStore{neighbors: []map[int][]uint64{nil}} // node 0 present but nil
+	g := &segGraphStore{neighbors: []map[int][]uint32{nil}} // node 0 present but nil
 
 	if got := g.getNeighborsRef(5, 0); got != nil { // id >= len(neighbors)
 		t.Errorf("getNeighborsRef(out-of-range) = %v, want nil", got)
@@ -230,7 +230,7 @@ func TestSegGraphStore_getNeighborsRef_Guard(t *testing.T) {
 		t.Errorf("getNeighborsRef(nil node) = %v, want nil", got)
 	}
 
-	g.neighbors[0] = map[int][]uint64{0: {7, 8}}
+	g.neighbors[0] = map[int][]uint32{0: {7, 8}}
 	got := g.getNeighborsRef(0, 0)
 	if len(got) != 2 || got[0] != 7 || got[1] != 8 {
 		t.Errorf("getNeighborsRef(valid) = %v, want [7 8]", got)
